implemented class structure
